Change comment in docs compilation script
The current comment says that the script watch for changes to .tex files. It actually listen on changes to .md.
Presumably the extension were changed to .md so Github will display the docs.

I'll admit that this is pedantry :-)

I came accross it while reading the docs and wondering how they are compiled.