fix: rm claim_rewards kwarg from  _transfer fn
It appears that adding an additional arg to the transfer/transferFrom
fns although set with a default, throws errors, and doesn't fit the
ERC20 spec. In any case, I've removed the claim_rewards kwarg from the
internal _tranfer function, and thereby set transfer calls to
automatically claim rewards (this seems like a better approach then
requiring users to submit two transactions, one to transfer and another
to claim rewards).

Also it should prevent an issue whereby, if there are two users left,
and one transfers their tokens to another, the other withdraws, the
first user is left rugged since they wouldn't be able to claim their
reward tokens (this is because technically the total supply = 0)