fix(test): side chain gauges expected emissions
Update of the test_emissions_against_expected test. This time it
appropriately verifies the behavior of the gauge, in that CRV is emitted
for the previous week at the previous week's rate, and that when
crossing the epoch period the rate is maintained until one week after
the epoch has passed.